Modification to create a LogManagement base class to be shared with telegram Youtube and Web scraping
