--- tests/monergism/procedures.txt
@@ -1,42 +1,0 @@
-
-1. Scrap the list of authors, topics and scriptures 
-
-    Run the methods in test_monergism_scrap_list.py 
-    The method test_get_topic_list() connect to https://www.monergism.com/topics and scrap the list of topics there 
-    The method test_get_speakers_list() connect to https://www.monergism.com/authors and scrap the list of authors there 
-    The method test_get_scripture_list() connect to https://www.monergism.com/scripture and scrap the list of topics there 
-
-
-2. Scrap for each author, topic and scripture the general informations 
-
-    This general informations consist of many things but the most important is that it 
-    contain the list of the pages of the authors. For exemple CH Spurgeon has 43 pages 
-
-    Run the method test_scrap_all_scripture_general_information() in the file scripture\test_mn_scrap_scripture_general_information.py 
-    to get the general informations about every scripture. Here the information are garthered for each book in th bible 
-    
-    Run the method test_scrap_all_author_general_information() in the file speaker\test_mn_scrap_speaker_general_information.py 
-    to get the general informations about every scripture. Here the information are garthered for each book in th bible 
-    
-    Run the method test_scrap_all_topic_general_information() in the file topic\test_mn_scrap_topic_general_information.py 
-    to get the general informations about every scripture. Here the information are garthered for each book in th bible 
-
-3. Scrap for each author, topic and scripture the works 
-
-    For each author, topic and scripture, the pages contain links with a text and an url. 
-    This works must be downloaded to be used later to download the content of those url. 
-
-    This part is time consuming because for each scripture, author or topic 
-    dozens of url request must be made. And there is dozens of topics and speakrs. 
-    For Spurgeon, there 43 request to be made 
-    
-
-    Run the method test_scrap_all_scripture_work() in the file scripture\test_mn_scrap_scripture_work.py 
-    to get the works of every scripture. 
-    Run the method test_scrap_all_topic_work() in the file topic\test_mn_scrap_topic_work.py
-    to get the works of every topic.
-
-    Run the method test_scrap_all_scripture_work() in the file speaker\test_mn_scrap_speaker_work.py
-    to get the works of every scripture. 
-
-4. Download the work of each author, topic and scripture
